画像共有サービスに使用するインタフェースを設計し直し
 * IMultimediaShareService から IMediaUploadService に変更
 * 複数枚のメディアをアップロード可能なサービスを考慮
 * URLのために確保する文字数を取得出来るようにした

+++ OpenTween/Api/TwitterConfiguration.cs
@@ -41,7 +41,7 @@ namespace OpenTween.Api
         public int CharactersReservedPerMedia { get; set; }
 
         [DataMember(Name = "photo_size_limit")]
-        public int PhotoSizeLimit { get; set; }
+        public long PhotoSizeLimit { get; set; }
 
         [DataMember(Name = "photo_sizes")]
         public TwitterMediaSizes PhotoSizes { get; set; }
